+++ src/shaders/glsl/line_gradient_sdf.fragment.glsl
@@ -4,7 +4,11 @@ uniform float u_mix;
 uniform lowp float u_lineatlas_width;
 
 in vec2 v_normal;
+#ifdef TAPER
+in vec2 v_width2;
+#else
 flat in vec2 v_width2;
+#endif
 in vec2 v_tex_a;
 in vec2 v_tex_b;
 in float v_gamma_scale;